prefs: Change count to 32 bit
This changes the count field in the prefs_t struct from 64 bit to 32
bit, since this should be enough for all meters in reality and is much
more performant and saves space.

+++ src/prefs.h
@@ -41,7 +41,7 @@ typedef struct {
     uint8_t max;
 
     // Actual counter (fixed point arithmetic 48 + 16?)
-    uint64_t count;
+    uint32_t count;
 
 } channel_prefs_t;
